requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
black==22.6.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Django==3.2.14
django-currentuser==0.5.3
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-flex-fields==0.9.8
django-extensions==3.2.0
jedi==0.18.1
Pygments==2.12.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
